Flood Map Feature Finished and Lightning feature fixed
Feature created to flood the map, dealing damage to player and enemies in water continuously while in contact. Tint overlay while submerged and warnings added for flood mechanic. Made small changes to lightning strike mechanic to block lightning bolts and damage when a structure is between player/enemies and bolt.